requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-telegram-bot==21.10
anthropic==0.64.0
supabase==2.15.3
APScheduler==3.11.0
python-dotenv==1.0.1
